.float()

+++ train.py
@@ -71,8 +71,8 @@ def train_epoch(model, training_data, crit, optimizer):
         n_total_correct += n_correct
         total_loss += loss.data[0]
 
-    train_loss = total_loss/n_total_words
-    train_accu = n_total_correct/n_total_words
+    train_loss = total_loss/n_total_words.float()
+    train_accu = n_total_correct/n_total_words.float()
     print("==>train_loss = \n",train_loss)
     print("==>train_accu = \n",train_accu)
     print("==>return train_epoch\n")
